specifies sizeof int across platforms

+++ add.c
@@ -5,8 +5,9 @@
  * 
  * Return: returns the sum of two integers..
  */
-int arbitrary_add(int num, int num3) {
-	int sum;
+uint64_t arbitrary_add(uint64_t num, uint64_t num3) {
+	uint64_t sum;
+
 	sum = num + num3;
 	return sum;
 }
